+++ src/msg/query_msg.rs
@@ -6,7 +6,7 @@ use cosmwasm_schema::{cw_serde, QueryResponses};
 #[derive(QueryResponses)]
 pub enum QueryMsg {
     #[returns(GetOrderResp)]
-    GetOrder { order_id: u128 },
+    GetOrder { order_id: u64 },
     #[returns(GetAllPricesResponse)]
     GetAllPrices {},
 }
