+++ src/tests/get_order_id_from_events.rs
@@ -1,11 +1,11 @@
 use super::*;
 
-pub fn get_order_id_from_events(events: &Vec<Event>) -> Option<u128> {
+pub fn get_order_id_from_events(events: &Vec<Event>) -> Option<u64> {
     let key = "order_id";
 
     for event in events {
         if let Some(attr) = event.attributes.iter().find(|attr| attr.key == key) {
-            if let Ok(id) = attr.value.parse::<u128>() {
+            if let Ok(id) = attr.value.parse::<u64>() {
                 return Some(id);
             }
         }
